Failing test for problem of parameterized generators
If we have a theory parameter of, say, Map<String, Integer>, and we have
several different generators that can produce instances of Maps, including
PropertiesGenerator, which is intended to produce Maps that are supposed
to contain only String keys and values, then when deciding what kinds of
generators should be used to produce values for the Map<String, Integer>
theory parameter, we need to allow the PropertiesGenerator to opt out,
or allow the framework to exclude the PropertiesGenerator, however clearest.

I'd like to minimize the burden on the generator writer.