For #51, introduce ForAll.seed
This allows the test writer to fix the seed material used by
junit-quickcheck' source of randomness. That way, if an unexpected
failure occurs for a theory and particular random data set, the
seed can be set for the next run in order to facilitate debugging.

junit-quickcheck hooks into the JUnit theories machinery through
ParameterSuppliers. As such, junit-quickcheck is fairly removed from the
process of creating tests, and listening for/reporting their results.
This presents a problem for junit-quickcheck in that there's not a really
standard way to report metadata about the theory and data set being
executed -- in particular, we want a way to report what seed value is
used in generating random values for a particular theory parameter.

In this implementation, the seed value for a particular theory parameter
is logged to an SLF4J logger, at debug level. This means that in order to
see the logged value, the test writer must supply an SLF4J binding on the
test class path.