For #32, shrinking failed property parameter lists
This means a new runner called JUnitQuickcheck. It operates on
methods marked as @Property. @ForAll is not required on parameters of
methods marked @Property.

JUnitQuickcheck invokes a @Property method Property.trials() times, and
so generates `trials` tuples of property parameters, whereas the
Theories runner in conjunction with junit-quickcheck's ForAll parameter
supplier generates all combinations of ForAll.sampleSize values for each
of the marked theory parameters.

Generators can tell the runner potential "shrinks" of a value. When a
property method fails, it is run again and again with shrinks of the
parameter set that caused the failure, until a set of values is found that
causes the property to succeed, or there are no more shrinks to try. The
"smallest" parameter set that caused failure is the one that is reported.

So far, the "built-in" generators allow shrinking of:

* Integral values (including BigInteger)
* Decimal values (incluing BigDecimal)
* Booleans
* Characters
* Strings
* Lists, sets, maps, arrays
* BitSets