Add SourceOfRandomness.nextElement()
Available with either array or collection arguments.

Preserving Items.choose for use internally, because it turned out to
be easier to write tests for internal classes doing a random choice of
elements via Items.choose than with SourceOfRandomness.nextElement.
Custom generators will now not have to resort to using an "internal"
method for this capability.