Correct stringifying of args/initial seeds
When raising an assertion error, in its messages we recently added the
initial seeds used to prime the random generator for generating property
values. Unfortunately, stringifying the Arrays.asList() of a long array
doesn't give the proper results. This change makes the stringifying of both
the property values and the initial seeds go through the proper overloads
of Arrays.toString().