Ensure ordering within a CompositeGenerator
Some tests rely on the ordering of items that the ServiceLoader emits.
The Maven plugin we use to generate the ServiceLoader provider config
files for generators cannot, and should not, enforce any particular
ordering of the contents of the generated files. Therefore, this change
enforces an ordering by class FQN of the generators fed to a
CompositeGenerator.

This problem became apparent when some tests failed unexpectedly on
Travis CI. We then discovered our tacit dependency on ordering within a
ServiceLoader provider config file.