For #100, add map/flatMap for generators
map and flatMap actually live as default methods on an interface Gen
extracted from Generator. This is a functional interface, so that you
can use lambdas to define your Gens.

These can be used within the definitions of Generators, especially in
combination with gen() (the method that allows a generator to get a handle
on a thing that helps it find other service-provided generators).