For #99, make original failure the cause of shrinking error
When a property fails, and shrinking kicks in, the reported assertion error
used to have the shrinking message, but the stack trace of the failure
that caused shrinking to occur. Now, the "top-level" assertion error is
the one raised by the shrinker, its stack trace is preserved, and the
assertion error that caused shrinking to happen in the first place is
embedded as the "cause" of the top-level error. That way, we have more
information about the failure.