Add submission from PR #198
It was necessary to correct a potential blocker that caused some
tests never to terminate. Basically, if a property parameter had a
`Map` type and for some reason a generator for `java.lang.Void` were
called upon to generate keys or values for the map, and the
`HashtableGenerator` were called on, you'd never find keys or values
that satisfied a `Hashtable`, b/c they'd all be `null`.

If you're explicit about `Void` as either a type parameter or a
wildcard bound on your `Map` property parameter, you'll still see the
interminable execution.  But, the likelihood of actually writing a
test like this is so low that we're willing to allow such interminability.
The behavior now is to prevent the `VoidGenerator` from being called upon
to produce values for `java.lang.Object` or unbounded wildcard types.